requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.1
requests==2.31.0
